# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Surveyvor App - Technical Manual",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get a JWT via given credentials",
                "description": "Returns a \\Illuminate\\Http\\JsonResponse with the generated token",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated User",
                "description": "Returns logged user data by authorization token passed",
                "operationId": "getLoggedUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Dereck"
                                        },
                                        "family_name": {
                                            "type": "string",
                                            "example": "Cepeda"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@surveyvor.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password1234"
                                        },
                                        "country_code": {
                                            "type": "string",
                                            "example": "ES"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "www.myimageprofile.com"
                                        },
                                        "telephone": {
                                            "type": "number",
                                            "example": "12345678"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create a new User",
                "description": "Returns the generated User",
                "operationId": "register",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Dereck"
                                    },
                                    "family_name": {
                                        "type": "string",
                                        "example": "Cepeda"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "test@surveyvor.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password1234"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "ES"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "www.myimageprofile.com"
                                    },
                                    "telephone": {
                                        "type": "number",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Dereck"
                                        },
                                        "family_name": {
                                            "type": "string",
                                            "example": "Cepeda"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@surveyvor.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password1234"
                                        },
                                        "country_code": {
                                            "type": "string",
                                            "example": "ES"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "www.myimageprofile.com"
                                        },
                                        "telephone": {
                                            "type": "number",
                                            "example": "12345678"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/all/{user_id}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get a User's public data",
                "description": "Returns a User's public data",
                "operationId": "getPublicDataFromUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Search by user_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Dereck"
                                        },
                                        "family_name": {
                                            "type": "string",
                                            "example": "Cepeda"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@surveyvor.com"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password1234"
                                        },
                                        "country_code": {
                                            "type": "string",
                                            "example": "ES"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "www.myimageprofile.com"
                                        },
                                        "telephone": {
                                            "type": "number",
                                            "example": "12345678"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/update/{user_id}": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update an existing User",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Search by user_id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Dereck"
                                    },
                                    "family_name": {
                                        "type": "string",
                                        "example": "Cepeda"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "ES"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "www.myimageprofile.com"
                                    },
                                    "telephone": {
                                        "type": "number",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}